requirements.txt
# ==============================================================================
# Banking77 Intent Classification – Dependencies
# ==============================================================================
# NOTE: On Google Colab, install unsloth first with:
#   pip install unsloth
# Then install the remaining dependencies below.
# ==============================================================================

unsloth
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
trl>=0.8.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.28.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
xformers
